update command system & fix bugs
fix #38

+++ src/main/java/org/soraworld/violet/util/FileUtils.java
@@ -27,28 +27,33 @@ public class FileUtils {
     }
 
     public static boolean zipArchivePath(Path source, Path target, Predicate<Path> filter) {
-        try {
-            ZipOutputStream zipOut = new ZipOutputStream(Files.newOutputStream(target));
-            Files.walk(source)
-                    .filter(filter)
-                    .filter(path -> !Files.isDirectory(path))
-                    .forEach(path -> {
-                        ZipEntry entry = new ZipEntry(source.relativize(path).toString());
-                        try {
-                            zipOut.putNextEntry(entry);
-                            Files.copy(path, zipOut);
-                            zipOut.closeEntry();
-                        } catch (IOException e) {
-                            e.printStackTrace();
-                        }
-                    });
-            zipOut.flush();
-            zipOut.close();
-            return true;
-        } catch (IOException e) {
-            e.printStackTrace();
-            return false;
+        if (Files.exists(source) && Files.isDirectory(source)) {
+            try {
+                Files.createDirectories(target.getParent());
+                Files.deleteIfExists(target);
+                Files.createFile(target);
+                ZipOutputStream zipOut = new ZipOutputStream(Files.newOutputStream(target));
+                Files.walk(source)
+                        .filter(filter)
+                        .filter(path -> !Files.isDirectory(path))
+                        .forEach(path -> {
+                            ZipEntry entry = new ZipEntry(source.relativize(path).toString());
+                            try {
+                                zipOut.putNextEntry(entry);
+                                Files.copy(path, zipOut);
+                                zipOut.closeEntry();
+                            } catch (IOException e) {
+                                e.printStackTrace();
+                            }
+                        });
+                zipOut.flush();
+                zipOut.close();
+                return true;
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
+        return false;
     }
 
 }
